% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_report_helpers.R
\name{aggregate_overall_statistics}
\alias{aggregate_overall_statistics}
\title{Aggregate overall analysis statistics for report}
\usage{
aggregate_overall_statistics(
  data,
  theme_set,
  consolidated = NULL,
  learning_context = NULL,
  config = NULL
)
}
\arguments{
\item{data}{tibble with all analysis columns}

\item{theme_set}{ThemeSet object}

\item{consolidated}{ConsolidatedCodes list (or NULL)}

\item{learning_context}{LearningContext object (or NULL)}

\item{config}{ThematicConfig object (or NULL)}
}
\value{
List of overall statistics
}
\description{
Aggregate overall analysis statistics for report
}
