% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_coding.R
\name{as_coding_results}
\alias{as_coding_results}
\title{Convert ProgressiveCodingState to legacy CodingResults format}
\usage{
as_coding_results(state)
}
\arguments{
\item{state}{ProgressiveCodingState}
}
\value{
CodingResults-compatible list
}
\description{
Convert ProgressiveCodingState to legacy CodingResults format
}
