% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_coding.R
\name{create_coding_state}
\alias{create_coding_state}
\title{Create a new progressive coding state}
\usage{
create_coding_state(learning_context = NULL, config_hash = NULL)
}
\arguments{
\item{learning_context}{LearningContext object (or NULL)}

\item{config_hash}{Hash of current config for resume compatibility}
}
\value{
ProgressiveCodingState S3 object
}
\description{
Create a new progressive coding state
}
