% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_correlations.R
\name{create_correlation_plot}
\alias{create_correlation_plot}
\title{Create correlation plot}
\usage{
create_correlation_plot(results, output_path)
}
\arguments{
\item{results}{CorrelationResults from calculate_correlations()}

\item{output_path}{File path for PNG output}
}
\description{
Create correlation plot
}
