% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/21_longitudinal.R
\name{.assign_period_labels}
\alias{.assign_period_labels}
\title{Convert a POSIXct vector into period labels}
\usage{
.assign_period_labels(timestamps, period_type)
}
\arguments{
\item{timestamps}{POSIXct vector}

\item{period_type}{One of \code{"daily"}, \code{"weekly"}, \code{"monthly"},
\code{"quarterly"}}
}
\value{
Character vector of period labels (same length as input)
}
\description{
Convert a POSIXct vector into period labels
}
\keyword{internal}
