% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_themes.R
\name{.cluster_cooccurrence_summary}
\alias{.cluster_cooccurrence_summary}
\title{Co-occurrence summary for a cluster}
\usage{
.cluster_cooccurrence_summary(cluster_leaves, codes, co_occurrence)
}
\description{
Co-occurrence summary for a cluster
}
\keyword{internal}
