% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_schemas.R
\name{.code_description_refresh_schema}
\alias{.code_description_refresh_schema}
\title{Schema for the code-description refresh response}
\usage{
.code_description_refresh_schema()
}
\description{
\preformatted{
  {
    "description": "<1-2 sentence refreshed description capturing the conceptual core>"
  }
}
}
\keyword{internal}
