% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_theme_data.R
\name{.code_from_codebook}
\alias{.code_from_codebook}
\title{Hydrate a Code S3 from a coding_state codebook entry}
\usage{
.code_from_codebook(key, coding_state)
}
\arguments{
\item{key}{Codebook key}

\item{coding_state}{ProgressiveCodingState}
}
\value{
Code S3 (stub if key not found in codebook)
}
\description{
Hydrate a Code S3 from a coding_state codebook entry
}
\keyword{internal}
