% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_comparison.R
\name{.code_jaccard}
\alias{.code_jaccard}
\title{Compute Jaccard similarity between two sets of strings}
\usage{
.code_jaccard(codes_a, codes_b)
}
\arguments{
\item{codes_a}{Character vector or semicolon-separated string}

\item{codes_b}{Character vector or semicolon-separated string}
}
\value{
Numeric between 0 and 1
}
\description{
Compute Jaccard similarity between two sets of strings
}
\keyword{internal}
