% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodology_decision_aid.R
\name{.config_wall_time_estimate}
\alias{.config_wall_time_estimate}
\title{Estimate wall-time for a run based on corpus size + mode}
\usage{
.config_wall_time_estimate(corpus_size, mode)
}
\description{
Estimate wall-time for a run based on corpus size + mode
}
\keyword{internal}
