% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_data_loading.R
\name{.default_column_mappings}
\alias{.default_column_mappings}
\title{Default column mappings for common platforms}
\usage{
.default_column_mappings()
}
\description{
Default column mappings for common platforms
}
\keyword{internal}
