% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_manuscript_learning.R
\name{.find_manuscript}
\alias{.find_manuscript}
\title{Find manuscript file in a study folder}
\usage{
.find_manuscript(folder, filenames)
}
\description{
Find manuscript file in a study folder
}
\keyword{internal}
