% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework_spec.R
\docType{data}
\name{.FRAMEWORK_SPEC_SCHEMA_VERSION}
\alias{.FRAMEWORK_SPEC_SCHEMA_VERSION}
\title{Current schema version for FrameworkSpec}
\format{
An object of class \code{character} of length 1.
}
\usage{
.FRAMEWORK_SPEC_SCHEMA_VERSION
}
\description{
Current schema version for FrameworkSpec
}
\keyword{internal}
