% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_human_verification.R
\name{.fuzzy_deduplicate_codes}
\alias{.fuzzy_deduplicate_codes}
\title{Deduplicate a list of codes by merging fuzzy near-duplicates}
\usage{
.fuzzy_deduplicate_codes(codes, threshold = 0.35)
}
\description{
Deduplicate a list of codes by merging fuzzy near-duplicates
}
\keyword{internal}
