% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_human_verification.R
\name{.interpret_kappa}
\alias{.interpret_kappa}
\title{Interpret kappa value using Landis & Koch (1977) scale}
\usage{
.interpret_kappa(kappa)
}
\description{
Interpret kappa value using Landis & Koch (1977) scale
}
\keyword{internal}
