% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live_tracking.R
\name{.live_atomic_write_json}
\alias{.live_atomic_write_json}
\title{Atomic JSON write (write-temp + rename)}
\usage{
.live_atomic_write_json(path, data)
}
\description{
Atomic JSON write (write-temp + rename)
}
\keyword{internal}
