% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_comparison.R
\name{.load_run_snapshot}
\alias{.load_run_snapshot}
\title{Load a single run's artifacts from disk}
\usage{
.load_run_snapshot(run_dir)
}
\arguments{
\item{run_dir}{Path to one run folder}
}
\value{
A RunSnapshot list
}
\description{
Load a single run's artifacts from disk
}
\keyword{internal}
