% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_human_verification.R
\name{.map_to_canonical}
\alias{.map_to_canonical}
\title{Map a set of codes to canonical codes via fuzzy matching}
\usage{
.map_to_canonical(codes, canonical, threshold = 0.35)
}
\description{
Map a set of codes to canonical codes via fuzzy matching
}
\keyword{internal}
