% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memos.R
\docType{data}
\name{.MEMO_SCHEMA_VERSION}
\alias{.MEMO_SCHEMA_VERSION}
\title{Current schema version for the Memo S3 class}
\format{
An object of class \code{character} of length 1.
}
\usage{
.MEMO_SCHEMA_VERSION
}
\description{
Current schema version for the Memo S3 class
}
\keyword{internal}
