% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_comparison.R
\name{.normalize_corr_var}
\alias{.normalize_corr_var}
\title{Normalize correlation variable names for matching across runs}
\usage{
.normalize_corr_var(x)
}
\description{
Normalize correlation variable names for matching across runs
}
\keyword{internal}
