% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_theme_data.R
\name{.one_legacy_subtheme}
\alias{.one_legacy_subtheme}
\title{Build one Subtheme S3 from a legacy list/row representation}
\usage{
.one_legacy_subtheme(s, all_code_names)
}
\description{
Build one Subtheme S3 from a legacy list/row representation
}
\keyword{internal}
