% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_manuscript_learning.R
\name{.parse_atlasti_codebook}
\alias{.parse_atlasti_codebook}
\title{Parse ATLAS.ti Excel codebook}
\usage{
.parse_atlasti_codebook(sheet_data)
}
\description{
Parse ATLAS.ti Excel codebook
}
\keyword{internal}
