% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_manuscript_learning.R
\name{.parse_maxqda_codebook}
\alias{.parse_maxqda_codebook}
\title{Parse MAXQDA Excel codebook}
\usage{
.parse_maxqda_codebook(sheet_data)
}
\description{
Parse MAXQDA Excel codebook
}
\keyword{internal}
