% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_manuscript_learning.R
\name{.parse_nvivo_codebook}
\alias{.parse_nvivo_codebook}
\title{Parse NVivo Excel codebook}
\usage{
.parse_nvivo_codebook(sheet_data)
}
\description{
Parse NVivo Excel codebook
}
\keyword{internal}
