% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote_provenance.R
\docType{data}
\name{.QUOTE_PROVENANCE_SCHEMA_VERSION}
\alias{.QUOTE_PROVENANCE_SCHEMA_VERSION}
\title{Current schema version for the quote provenance object}
\format{
An object of class \code{character} of length 1.
}
\usage{
.QUOTE_PROVENANCE_SCHEMA_VERSION
}
\description{
Current schema version for the quote provenance object
}
\keyword{internal}
