% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_json_utils.R
\name{.repair_close_brackets}
\alias{.repair_close_brackets}
\title{Close unclosed brackets and braces}
\usage{
.repair_close_brackets(json)
}
\description{
Close unclosed brackets and braces
}
\keyword{internal}
