% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_state.R
\docType{data}
\name{.RUN_METADATA_SCHEMA_VERSION}
\alias{.RUN_METADATA_SCHEMA_VERSION}
\title{Schema version for run_metadata.json (T1.4 / T1.5)}
\format{
An object of class \code{character} of length 1.
}
\usage{
.RUN_METADATA_SCHEMA_VERSION
}
\description{
Consumers (replay_run, summarize_audit_log) read this to gate
back-compat reads. Bumped when the metadata schema changes
incompatibly.
}
\keyword{internal}
