% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_comparison.R
\docType{data}
\name{.SCHEMA_VERSION}
\alias{.SCHEMA_VERSION}
\title{Current output schema version this package writes}
\format{
An object of class \code{character} of length 1.
}
\usage{
.SCHEMA_VERSION
}
\description{
Current output schema version this package writes
}
\keyword{internal}
