% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_theme_data.R
\name{.subtheme_name_desc_pairs}
\alias{.subtheme_name_desc_pairs}
\title{Subtheme name+description pairs for tibble serialization}
\usage{
.subtheme_name_desc_pairs(theme)
}
\description{
Subtheme name+description pairs for tibble serialization
}
\keyword{internal}
