% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_theme_data.R
\name{.subtheme_names_no_virtual}
\alias{.subtheme_names_no_virtual}
\title{Subtheme names omitting virtual (NA-named) wrappers}
\usage{
.subtheme_names_no_virtual(theme)
}
\description{
Subtheme names omitting virtual (NA-named) wrappers
}
\keyword{internal}
