% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_theme_data.R
\docType{data}
\name{.THEME_DEFAULTS}
\alias{.THEME_DEFAULTS}
\title{Default values for optional theme fields}
\format{
An object of class \code{list} of length 8.
}
\usage{
.THEME_DEFAULTS
}
\description{
Default values for optional theme fields
}
\keyword{datasets}
