% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_theme_data.R
\docType{data}
\name{.THEME_REQUIRED_FIELDS}
\alias{.THEME_REQUIRED_FIELDS}
\title{Required fields for each theme within a ThemeSet}
\format{
An object of class \code{character} of length 3.
}
\usage{
.THEME_REQUIRED_FIELDS
}
\description{
Required fields for each theme within a ThemeSet
}
\keyword{datasets}
