% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transparency_report.R
\name{.tr_reflexivity_completeness}
\alias{.tr_reflexivity_completeness}
\title{Score reflexivity scaffold completeness (0/3 .. 3/3)}
\usage{
.tr_reflexivity_completeness(study)
}
\description{
Score reflexivity scaffold completeness (0/3 .. 3/3)
}
\keyword{internal}
