% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transparency_report.R
\name{.tr_render_html}
\alias{.tr_render_html}
\title{Render the transparency report as self-contained HTML}
\usage{
.tr_render_html(rd)
}
\description{
Render the transparency report as self-contained HTML
}
\keyword{internal}
