% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework_spec.R
\docType{data}
\name{.VALID_EPISTEMIC_STANCES}
\alias{.VALID_EPISTEMIC_STANCES}
\title{Valid epistemic-stance enum values}
\format{
An object of class \code{character} of length 3.
}
\usage{
.VALID_EPISTEMIC_STANCES
}
\description{
Valid epistemic-stance enum values
}
\keyword{internal}
