% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provocateur.R
\docType{data}
\name{.VALID_PROVOCATION_CATEGORIES}
\alias{.VALID_PROVOCATION_CATEGORIES}
\title{Valid provocation category names}
\format{
An object of class \code{character} of length 5.
}
\usage{
.VALID_PROVOCATION_CATEGORIES
}
\description{
Valid provocation category names
}
\keyword{internal}
