% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote_provenance.R
\docType{data}
\name{.VALID_QUOTE_CITATION_SOURCES}
\alias{.VALID_QUOTE_CITATION_SOURCES}
\title{Valid citation sources (where the quote came from)}
\format{
An object of class \code{character} of length 4.
}
\usage{
.VALID_QUOTE_CITATION_SOURCES
}
\description{
Valid citation sources (where the quote came from)
}
\keyword{internal}
