% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_report.R
\name{export_results}
\alias{export_results}
\title{Export all analysis results to files}
\usage{
export_results(
  data,
  theme_set,
  correlations_df,
  insights,
  consolidated,
  output_dir
)
}
\arguments{
\item{data}{tibble with all analysis columns}

\item{theme_set}{ThemeSet object}

\item{correlations_df}{Correlations tibble}

\item{insights}{Insights list}

\item{consolidated}{ConsolidatedCodes list}

\item{output_dir}{Output directory path}
}
\value{
List of export file paths
}
\description{
Export all analysis results to files
}
