% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_report_helpers.R
\name{generate_downloads_section}
\alias{generate_downloads_section}
\title{Generate downloads appendix section}
\usage{
generate_downloads_section(export_files, theme_stats)
}
\arguments{
\item{export_files}{List of export file paths}

\item{theme_stats}{Per-theme statistics list}
}
\value{
R Markdown string
}
\description{
Generate downloads appendix section
}
