% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_run_id}
\alias{generate_run_id}
\title{Generate a unique run ID based on timestamp}
\usage{
generate_run_id()
}
\value{
Character string like "run_2026-02-23_143052"
}
\description{
Generate a unique run ID based on timestamp
}
