% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_coding.R
\name{get_analytic_sample}
\alias{get_analytic_sample}
\title{Get the analytic sample (entries that received at least one code)}
\usage{
get_analytic_sample(state, data)
}
\arguments{
\item{state}{ProgressiveCodingState}

\item{data}{Full data tibble}
}
\value{
Filtered tibble containing only coded entries
}
\description{
Get the analytic sample (entries that received at least one code)
}
