% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_report_helpers.R
\name{get_emotion_interpretation}
\alias{get_emotion_interpretation}
\title{Get interpretation text for an emotion}
\usage{
get_emotion_interpretation(emotion)
}
\arguments{
\item{emotion}{Emotion name string}
}
\value{
Interpretation string
}
\description{
Get interpretation text for an emotion
}
