% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_checkpoint.R
\name{hash_config}
\alias{hash_config}
\title{Compute a hash of a config file for change detection}
\usage{
hash_config(config_path)
}
\arguments{
\item{config_path}{Path to YAML config file}
}
\value{
Character hash string
}
\description{
Compute a hash of a config file for change detection
}
