% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_manuscript_learning.R
\name{load_previous_studies}
\alias{load_previous_studies}
\title{Load all previous studies from a base directory}
\usage{
load_previous_studies(base_dir, config = list())
}
\arguments{
\item{base_dir}{Absolute path to manual analyses root}

\item{config}{Learning config section from YAML}
}
\value{
PreviousStudies S3 object
}
\description{
Load all previous studies from a base directory
}
