% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_anchor_id}
\alias{make_anchor_id}
\title{Create an HTML anchor ID from a string}
\usage{
make_anchor_id(name)
}
\arguments{
\item{name}{Input string}
}
\value{
Anchor-safe string
}
\description{
Create an HTML anchor ID from a string
}
