% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_safe_filename}
\alias{make_safe_filename}
\title{Create a safe filename from a string}
\usage{
make_safe_filename(name)
}
\arguments{
\item{name}{Input string}
}
\value{
Filesystem-safe string
}
\description{
Create a safe filename from a string
}
