% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null-coalesce}
\alias{null-coalesce}
\alias{\%||\%}
\title{Null-coalescing operator (re-export from rlang)}
\usage{
x \%||\% y
}
\description{
Null-coalescing operator (re-export from rlang)
}
\keyword{internal}
