% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_checkpoint.R
\name{print.CheckpointManager}
\alias{print.CheckpointManager}
\title{Print method for CheckpointManager}
\usage{
\method{print}{CheckpointManager}(x, ...)
}
\arguments{
\item{x}{CheckpointManager object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for CheckpointManager
}
