% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_comparison.R
\name{print.ComparisonResult}
\alias{print.ComparisonResult}
\title{Print method for ComparisonResult}
\usage{
\method{print}{ComparisonResult}(x, ...)
}
\arguments{
\item{x}{ComparisonResult object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for ComparisonResult
}
