% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_coverage.R
\name{print.CorpusCoverage}
\alias{print.CorpusCoverage}
\title{Print method for CorpusCoverage}
\usage{
\method{print}{CorpusCoverage}(x, ...)
}
\arguments{
\item{x}{A CorpusCoverage object}

\item{...}{Ignored}
}
\description{
Print method for CorpusCoverage
}
