% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework_spec.R
\name{print.FrameworkSpec}
\alias{print.FrameworkSpec}
\title{Print method for FrameworkSpec}
\usage{
\method{print}{FrameworkSpec}(x, ...)
}
\arguments{
\item{x}{A FrameworkSpec object}

\item{...}{Ignored}
}
\description{
Print method for FrameworkSpec
}
