% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live_tracking.R
\name{print.LiveTracker}
\alias{print.LiveTracker}
\title{Print method for LiveTracker}
\usage{
\method{print}{LiveTracker}(x, ...)
}
\arguments{
\item{x}{LiveTracker object}

\item{...}{ignored}
}
\description{
Print method for LiveTracker
}
