% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provocateur.R
\name{print.Provocation}
\alias{print.Provocation}
\title{Print method for Provocation}
\usage{
\method{print}{Provocation}(x, ...)
}
\arguments{
\item{x}{A Provocation}

\item{...}{Ignored}
}
\description{
Print method for Provocation
}
