% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode1_orchestrator.R
\name{print.ProvocationCoverage}
\alias{print.ProvocationCoverage}
\title{Print method for ProvocationCoverage}
\usage{
\method{print}{ProvocationCoverage}(x, ...)
}
\arguments{
\item{x}{A ProvocationCoverage object}

\item{...}{Ignored}
}
\description{
Print method for ProvocationCoverage
}
