% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote_provenance.R
\name{print.QuoteProvenance}
\alias{print.QuoteProvenance}
\title{Print method for QuoteProvenance}
\usage{
\method{print}{QuoteProvenance}(x, ...)
}
\arguments{
\item{x}{A QuoteProvenance object}

\item{...}{Ignored}
}
\description{
Print method for QuoteProvenance
}
