% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provocateur.R
\name{print.ResearcherReflectionLog}
\alias{print.ResearcherReflectionLog}
\title{Print method for ResearcherReflectionLog}
\usage{
\method{print}{ResearcherReflectionLog}(x, ...)
}
\arguments{
\item{x}{A ResearcherReflectionLog}

\item{...}{Ignored}
}
\description{
Print method for ResearcherReflectionLog
}
