% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_theme_data.R
\name{print.Subtheme}
\alias{print.Subtheme}
\title{Print method for Subtheme}
\usage{
\method{print}{Subtheme}(x, ...)
}
\arguments{
\item{x}{Subtheme object}

\item{...}{ignored}
}
\description{
Print method for Subtheme
}
