% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_config.R
\name{print.ThematicConfig}
\alias{print.ThematicConfig}
\title{Print method for ThematicConfig}
\usage{
\method{print}{ThematicConfig}(x, ...)
}
\arguments{
\item{x}{ThematicConfig object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for ThematicConfig
}
