% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_theme_data.R
\name{print.ThemeSet}
\alias{print.ThemeSet}
\title{Print method for ThemeSet}
\usage{
\method{print}{ThemeSet}(x, ...)
}
\arguments{
\item{x}{ThemeSet object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for ThemeSet
}
