% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_checkpoint.R
\name{save_checkpoint}
\alias{save_checkpoint}
\title{Save checkpoint at a given pipeline step}
\usage{
save_checkpoint(manager, step_name, data)
}
\arguments{
\item{manager}{CheckpointManager object}

\item{step_name}{Step identifier (e.g., "sentiment_done")}

\item{data}{The data/results to save}
}
\description{
Save checkpoint at a given pipeline step
}
