% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_data_loading.R
\name{standardize_data}
\alias{standardize_data}
\title{Standardize data to common schema}
\usage{
standardize_data(data, column_map)
}
\arguments{
\item{data}{Raw tibble}

\item{column_map}{Result of detect_columns()}
}
\value{
tibble with std_id, std_text, std_author, std_timestamp, + original metrics
}
\description{
Standardize data to common schema
}
