% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_theme_data.R
\name{subtheme_code_keys}
\alias{subtheme_code_keys}
\title{Code keys within a Subtheme}
\usage{
subtheme_code_keys(subtheme)
}
\arguments{
\item{subtheme}{Subtheme S3}
}
\value{
Character vector
}
\description{
Code keys within a Subtheme
}
