% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_theme_data.R
\name{subtheme_code_names}
\alias{subtheme_code_names}
\title{Code names (display) within a Subtheme}
\usage{
subtheme_code_names(subtheme)
}
\arguments{
\item{subtheme}{Subtheme S3}
}
\value{
Character vector
}
\description{
Code names (display) within a Subtheme
}
