% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_theme_data.R
\name{subtheme_n_codes}
\alias{subtheme_n_codes}
\title{Number of codes in a Subtheme}
\usage{
subtheme_n_codes(subtheme)
}
\arguments{
\item{subtheme}{Subtheme S3}
}
\value{
Integer
}
\description{
Number of codes in a Subtheme
}
