% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_theme_data.R
\name{theme_code_keys}
\alias{theme_code_keys}
\title{Flatten code keys across all subthemes of a theme}
\usage{
theme_code_keys(theme)
}
\arguments{
\item{theme}{A theme list}
}
\value{
Character vector of code KEYS
}
\description{
Flatten code keys across all subthemes of a theme
}
