% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_theme_data.R
\name{theme_codes}
\alias{theme_codes}
\title{Flatten code names across all subthemes of a theme (back-compat with codes_included)}
\usage{
theme_codes(theme)
}
\arguments{
\item{theme}{A theme list}
}
\value{
Character vector of code NAMES
}
\description{
Flatten code names across all subthemes of a theme (back-compat with codes_included)
}
