% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_theme_data.R
\name{theme_names}
\alias{theme_names}
\title{Extract theme names from ThemeSet}
\usage{
theme_names(theme_set)
}
\arguments{
\item{theme_set}{ThemeSet object}
}
\value{
Character vector of theme names
}
\description{
Extract theme names from ThemeSet
}
