% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_theme_data.R
\name{theme_segments}
\alias{theme_segments}
\title{Flatten coded_segments across all codes of a theme}
\usage{
theme_segments(theme)
}
\arguments{
\item{theme}{A theme list}
}
\value{
Flat list of segment records
}
\description{
Flatten coded_segments across all codes of a theme
}
