% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{truncate_text}
\alias{truncate_text}
\title{Truncate text to specified length with ellipsis}
\usage{
truncate_text(text, max_length = 200)
}
\arguments{
\item{text}{Input string}

\item{max_length}{Maximum characters}
}
\value{
Truncated string
}
\description{
Truncate text to specified length with ellipsis
}
