% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{validate_class}
\alias{validate_class}
\title{Validate that an object inherits from a given class}
\usage{
validate_class(x, cls, caller = NULL)
}
\arguments{
\item{x}{Object to check}

\item{cls}{Expected class name}

\item{caller}{Name of the calling function (for error messages)}
}
\description{
Validate that an object inherits from a given class
}
\keyword{internal}
