% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_sentiment.R
\name{.assign_sentiment_results}
\alias{.assign_sentiment_results}
\title{Assign parsed sentiment results back to data}
\usage{
.assign_sentiment_results(data, results_data, batch_rows)
}
\description{
Assign parsed sentiment results back to data
}
\keyword{internal}
