% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_themes.R
\name{.coalesce_virtual_subtheme_groups}
\alias{.coalesce_virtual_subtheme_groups}
\title{Combine adjacent NA-named subtheme groups into one virtual group}
\usage{
.coalesce_virtual_subtheme_groups(groups)
}
\description{
Combine adjacent NA-named subtheme groups into one virtual group
}
\keyword{internal}
