% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_coverage.R
\docType{data}
\name{.CORPUS_COVERAGE_SCHEMA_VERSION}
\alias{.CORPUS_COVERAGE_SCHEMA_VERSION}
\title{Current schema version for the CorpusCoverage object}
\format{
An object of class \code{character} of length 1.
}
\usage{
.CORPUS_COVERAGE_SCHEMA_VERSION
}
\description{
Current schema version for the CorpusCoverage object
}
\keyword{internal}
