% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provocateur.R
\docType{data}
\name{.RESEARCHER_REFLECTION_LOG_SCHEMA_VERSION}
\alias{.RESEARCHER_REFLECTION_LOG_SCHEMA_VERSION}
\title{Current ResearcherReflectionLog schema version}
\format{
An object of class \code{character} of length 1.
}
\usage{
.RESEARCHER_REFLECTION_LOG_SCHEMA_VERSION
}
\description{
Current ResearcherReflectionLog schema version
}
\keyword{internal}
