% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_theme_data.R
\name{theme_code_objects}
\alias{theme_code_objects}
\title{Flatten Code S3 objects across all subthemes of a theme}
\usage{
theme_code_objects(theme)
}
\arguments{
\item{theme}{A theme list (one element of theme_set$themes)}
}
\value{
List of Code S3 objects
}
\description{
Flatten Code S3 objects across all subthemes of a theme
}
