% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_theme_data.R
\name{theme_n_subthemes}
\alias{theme_n_subthemes}
\title{Number of subthemes in a theme (excludes virtual single-subtheme wrappers)}
\usage{
theme_n_subthemes(theme)
}
\arguments{
\item{theme}{A theme list}
}
\value{
Integer
}
\description{
Number of subthemes in a theme (excludes virtual single-subtheme wrappers)
}
