% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_theme_data.R
\name{theme_set_to_tibble}
\alias{theme_set_to_tibble}
\title{Convert ThemeSet to tibble for export/inspection}
\usage{
theme_set_to_tibble(theme_set)
}
\arguments{
\item{theme_set}{ThemeSet object}
}
\value{
tibble with one row per theme
}
\description{
Convert ThemeSet to tibble for export/inspection
}
